requirements.txt
PyGithub==1.59.1
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0
slack-sdk==3.21.3
atlassian-python-api==3.41.3
python-dateutil==2.8.2
schedule==1.2.0